requirements.txt
datadreamer.dev<=0.35.0
openai<=1.35.5
datasets<=2.19.0
sentence-transformers<=2.3.0
transformers<=4.40.1
accelerate<=0.29.3
matplotlib<=3.9.0
seaborn<=0.13.2
playwright<=1.44.0
beautifulsoup4<=4.12.3
PyLaTeX<=1.4.2
pdf2image<=1.17.0
graphviz<=0.20.3
pandas<=2.2.2
scipy<=1.14.0
networkx<=3.3
geopandas<=1.0.0
shapely<=2.0.4
altair<=5.3.0
vl-convert-python<=1.5.0
plotly<=5.22.0
kaleido<=0.2.1.post1
vega_datasets<=0.9.0
altair-saver<=0.5.0
vl-convert-python<=1.5.0
weasyprint<=62.3
openpyxl<=3.1.5
dataframe_image<=0.2.4
reportlab<=4.2.2
webdriver_manager<=4.0.1
docx2pdf<=0.1.8
python-docx<=1.1.2
ordered-set<=4.1.0
squarify<=0.4.4
markdown2<=2.5.0
bokeh<=3.5.1
rdkit<=2024.3.5
ezdxf<=1.3.3
music21<=9.1.0
svglib<=1.5.1
threadpoolctl>=3.1.0
setfit<=1.0.3
huggingface_hub<=0.23.4
peft<=0.11.1
datasets<=2.19.0
httpx<=0.27.2 